+++ qiskit/pulse/utils.py
@@ -22,14 +22,15 @@ from typing import List, Optional
 import numpy as np
 
 from .channels import AcquireChannel, MemorySlot
-from .cmd_def import CmdDef
+from .qasm_def import QasmToSchedDef
 from .commands import Acquire, AcquireInstruction
 from .exceptions import PulseError
 from .interfaces import ScheduleComponent
 from .schedule import Schedule
 
 
-def align_measures(schedules: List[ScheduleComponent], cmd_def: CmdDef, cal_gate: str = 'u3',
+def align_measures(schedules: List[ScheduleComponent], qasm_def: QasmToSchedDef,
+                   cal_gate: str = 'u3',
                    max_calibration_duration: Optional[int] = None,
                    align_time: Optional[int] = None) -> Schedule:
     """Return new schedules where measurements occur at the same physical time. Minimum measurement
@@ -40,9 +41,9 @@ def align_measures(schedules: List[ScheduleComponent], cmd_def: CmdDef, cal_gate
 
     Args:
         schedules: Collection of schedules to be aligned together
-        cmd_def: Command definition list
+        qasm_def: Command definition list
         cal_gate: The name of the gate to inspect for the calibration time
-        max_calibration_duration: If provided, cmd_def and cal_gate will be ignored
+        max_calibration_duration: If provided, qasm_def and cal_gate will be ignored
         align_time: If provided, this will be used as final align time.
     Returns:
         Schedule
@@ -57,9 +58,9 @@ def align_measures(schedules: List[ScheduleComponent], cmd_def: CmdDef, cal_gate
         # Need time to allow for calibration pulses to be played for result classification
         if max_calibration_duration is None:
             max_calibration_duration = 0
-            for qubits in cmd_def.cmd_qubits(cal_gate):
-                cmd = cmd_def.get(cal_gate, qubits, np.pi, 0, np.pi)
-                max_calibration_duration = max(cmd.duration, max_calibration_duration)
+            for qubits in qasm_def.gate_qubits(cal_gate):
+                sched = qasm_def.get(cal_gate, qubits, np.pi, 0, np.pi)
+                max_calibration_duration = max(sched.duration, max_calibration_duration)
 
         # Schedule the acquires to be either at the end of the needed calibration time, or when the
         # last acquire is scheduled, whichever comes later