+++ CHANGELOG.md
@@ -71,6 +71,7 @@ The format is based on [Keep a Changelog].
 -   Replaces LegacySwap by faster, more stable StochasticSwap pass (\#2672)
 -   Uses level 1 by default as transpiler optimization level (\#2672)
 -   Change Snapshot signature to match simulator.snapshot (\#2592)
+-   Rename CmdDef to QasmToSchedDef (\#2681)
 
 ### Removed